requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
pymupdf
sentence-transformers
faiss-cpu
openai
python-dotenv
numpy
